enhanced modularization
new files include better modularization, new util file
